+++ packages/icons/src/names.ts
@@ -533,5 +533,6 @@ export const names: IconName[] = [
   'politicsCandidate',
   'ballot',
   'rain',
-  'airdropParachute'
+  'airdropParachute',
+  'birthcertificate'
 ];