+++ packages/illustrations/src/__generated__/pictogram/types/PictogramName.ts
@@ -248,6 +248,7 @@ export type PictogramName =
   | 'commerceCheckout'
   | 'commerceInvoice'
   | 'commerceNavigation'
+  | 'commodities'
   | 'completeQuiz'
   | 'complianceNavigation'
   | 'congratulations'