+++ packages/illustrations/src/__generated__/spotSquare/data/names.ts
@@ -158,6 +158,7 @@ const names: SpotSquareName[] = [
   'giftBoxCrypto',
   'gifting',
   'globalTransactions',
+  'goldSilverFutures',
   'guideBullCase',
   'guideCryptoBeginner',
   'guideFiveThings',