+++ packages/illustrations/src/__generated__/spotSquare/data/versionMap.ts
@@ -261,8 +261,9 @@ const versionMap: Record<SpotSquareName, number> = {
   baseRewardTrophyStars: 0,
   coinbaseUnlockOffers: 0,
   baseCreatorCoin: 1,
-  cryptoEconomyArrows: 1,
+  cryptoEconomyArrows: 2,
   baseQuickBuy: 0,
+  goldSilverFutures: 0,
 };
 
 export default versionMap;