+++ packages/illustrations/src/__generated__/spotSquare/types/SpotSquareName.ts
@@ -152,6 +152,7 @@ export type SpotSquareName =
   | 'giftBoxCrypto'
   | 'gifting'
   | 'globalTransactions'
+  | 'goldSilverFutures'
   | 'guideBullCase'
   | 'guideCryptoBeginner'
   | 'guideFiveThings'