+++ packages/icons/CHANGELOG.md
@@ -8,6 +8,21 @@ All notable changes to this project will be documented in this file.
 
 <!-- template-start -->
 
+## 5.10.0 (1/29/2026 PST)
+
+#### 🚀 Updates
+
+- Feat: Publish icons 2026-01-29. [[#342](https://github.com/coinbase/cds/pull/342)]
+
+##### ⭐️ Added (1)
+
+- birthcertificate
+
+##### ⭐️ Updated (1)
+
+- smartContract
+- pencil
+
 ## 5.9.0 (12/22/2025 PST)
 
 #### 🚀 Updates